refactor seadragon-svg  - add import, export; get rid of require
